# backend/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: TODO
paths:
  /signup:
    post:
      summary: TODO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/signup"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userCredentials"
  /login:
    post:
      summary: TODO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/login"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/userCredentials"
                  - $ref: "#/components/schemas/TotpRequired"
  /totp/enable:
    post:
      summary: TODO
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnableTotp"
  /totp/enable/verify:
    patch:
      summary: TODO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableTotpVerify"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryCodes"
  /totp/disable:
    patch:
      summary: TODO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Totp"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                type: boolean
                example: true
  /totp/authenticate:
    post:
      summary: TODO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableTotpVerify"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userCredentials"
  /totp/recovery:
    post:
      summary: TODO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryCode"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userCredentials"
  /getCalendars:
    get:
      summary: TODO
      security:
        - BearerAuth: []
      responses:
        "200":
          description: TODO
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Calendar"
  /addCalendar:
    post:
      summary: TODO
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarNoId"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
  /calendar/edit/{calendar_id}:
    patch:
      summary: TODO
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: calendar_id
          description: TODO
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
            example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEdit"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
  /calendar/delete/{calendar_id}:
    delete:
      summary: TODO
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: calendar_id
          description: TODO
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
            example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
          required: true
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                type: boolean
                example: true
  /addEvent:
    post:
      summary: TODO
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddEvent"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /getEvents:
    get:
      summary: TODO
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: start_time
          schema:
            type: string
            format: date-time
            example: TODO
          required: true
          description: TODO
        - in: query
          name: end_time
          schema:
            type: string
            format: date-time
            example: TODO
          required: true
          description: TODO
      responses:
        "200":
          description: TODO
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
  /event/edit/{event_id}:
    patch:
      summary: TODO
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event_id
          description: TODO
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
            example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventEdit"
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /event/delete/{event_id}:
    delete:
      summary: TODO
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event_id
          description: TODO
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
            example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
          required: true
      responses:
        "201":
          description: TODO
          content:
            application/json:
              schema:
                type: boolean
                example: true
components:
  securitySchemas:
    BearerAuth:
      type: http
      schema: bearer
      bearerFormat: JWT
  schemas:
    Calendar:
      type: object
      required: [id, name, owner_id]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
          example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
        name:
          type: string
          example: "calendar 1"
        owner_id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
          example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
    CalendarNoId:
      type: object
      required: [name, owner_id]
      properties:
        name:
          type: string
          example: "calendar 1"
        owner_id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
          example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
    CalendarEdit:
      type: object
      properties:
        name:
          type: string
          example: "edited calendar 1"
    signup:
      type: object
      required: [name, password, password_confirmation]
      properties:
        name:
          type: string
          example: "user 1"
        password:
          type: string
          example: "password 1"
        password_confirmation:
          type: string
          example: "password 1"
    login:
      type: object
      required: [name, password]
      properties:
        name:
          type: string
          example: "user1"
        password:
          type: string
          example: "password1"
    userCredentials:
      type: object
      required: [name, JWT]
      properties:
        name:
          type: string
        JWT:
          type: string
    Event:
      type: object
      required: [id, name, calendar_id, start_time, end_time]
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
          example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
        name:
          type: string
          example: "event 1"
        calendar_id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
          example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
        start_time:
          type: string
          format: date-time
          example: 2022-01-02T16:00:00+07:00
        end_time:
          type: string
          format: date-time
          example: 2022-01-02T16:00:00+07:00
    AddEvent:
      type: object
      required: [name, calendar_id, start_time, end_time]
      properties:
        name:
          type: string
          example: "event 1"
        calendar_id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
          example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
        start_time:
          type: string
          format: date-time
          example: 2022-01-02T16:00:00+07:00
        end_time:
          type: string
          format: date-time
          example: 2022-01-02T16:00:00+07:00
    EventEdit:
      type: object
      properties:
        calendar_id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
          example: "019b7abf-0d2a-7240-bc33-68fc0a8a1118"
        name:
          type: string
          example: "edited event 1"
        start_time:
          type: string
          format: date-time
          example: 2022-01-02T16:00:00+07:00
        end_time:
          type: string
          format: date-time
          example: 2022-01-02T16:00:00+07:00
    Totp:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          example: 123456
    TotpRequired:
      type: object
      required: [verification_token]
      properties:
        verification_token:
          type: string
    EnableTotp:
      type: object
      required: [verification_token]
      properties:
        verification_token:
          type: string
    EnableTotpVerify:
      type: object
      required: [code, verification_token]
      properties:
        code:
          type: integer
          example: 123456
        verification_token:
          type: string
    RecoveryCode:
      type: object
      required: [recovery_code, verification_token]
      properties:
        recovery_code:
          type: string
          example: "AAAA-BBBB-CCCC"
        verification_token:
          type: string
    RecoveryCodes:
      type: object
      required: [recovery_codes]
      properties:
        recovery_codes:
          type: array
          minItems: 10
          maxItems: 10
          items:
            type: string
            example: "AAAA-BBBB-CCCC"