# backend/internal/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/google/uuid"
	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// AddEvent defines model for AddEvent.
type AddEvent struct {
	CalendarId uuid.UUID `json:"calendar_id"`
	EndTime    time.Time `json:"end_time"`
	Name       string    `json:"name"`
	StartTime  time.Time `json:"start_time"`
}

// Calendar defines model for Calendar.
type Calendar struct {
	Id      uuid.UUID `json:"id"`
	Name    string    `json:"name"`
	OwnerId uuid.UUID `json:"owner_id"`
}

// CalendarEdit defines model for CalendarEdit.
type CalendarEdit struct {
	Name *string `json:"name,omitempty"`
}

// CalendarNoId defines model for CalendarNoId.
type CalendarNoId struct {
	Name    string    `json:"name"`
	OwnerId uuid.UUID `json:"owner_id"`
}

// EnableTotp defines model for EnableTotp.
type EnableTotp struct {
	VerificationToken string `json:"verification_token"`
}

// EnableTotpVerify defines model for EnableTotpVerify.
type EnableTotpVerify struct {
	Code              int    `json:"code"`
	VerificationToken string `json:"verification_token"`
}

// Event defines model for Event.
type Event struct {
	CalendarId uuid.UUID `json:"calendar_id"`
	EndTime    time.Time `json:"end_time"`
	Id         uuid.UUID `json:"id"`
	Name       string    `json:"name"`
	StartTime  time.Time `json:"start_time"`
}

// EventEdit defines model for EventEdit.
type EventEdit struct {
	CalendarId *uuid.UUID `json:"calendar_id,omitempty"`
	EndTime    *time.Time `json:"end_time,omitempty"`
	Name       *string    `json:"name,omitempty"`
	StartTime  *time.Time `json:"start_time,omitempty"`
}

// RecoveryCode defines model for RecoveryCode.
type RecoveryCode struct {
	RecoveryCode      string `json:"recovery_code"`
	VerificationToken string `json:"verification_token"`
}

// RecoveryCodes defines model for RecoveryCodes.
type RecoveryCodes struct {
	RecoveryCodes []string `json:"recovery_codes"`
}

// Totp defines model for Totp.
type Totp struct {
	Code int `json:"code"`
}

// TotpRequired defines model for TotpRequired.
type TotpRequired struct {
	VerificationToken string `json:"verification_token"`
}

// Login defines model for login.
type Login struct {
	Name     string `json:"name"`
	Password string `json:"password"`
}

// Signup defines model for signup.
type Signup struct {
	Name                 string `json:"name"`
	Password             string `json:"password"`
	PasswordConfirmation string `json:"password_confirmation"`
}

// UserCredentials defines model for userCredentials.
type UserCredentials struct {
	JWT  string `json:"JWT"`
	Name string `json:"name"`
}

// GetGetEventsParams defines parameters for GetGetEvents.
type GetGetEventsParams struct {
	// StartTime TODO
	StartTime time.Time `form:"start_time" json:"start_time"`

	// EndTime TODO
	EndTime time.Time `form:"end_time" json:"end_time"`
}

// PostAddCalendarJSONRequestBody defines body for PostAddCalendar for application/json ContentType.
type PostAddCalendarJSONRequestBody = CalendarNoId

// PostAddEventJSONRequestBody defines body for PostAddEvent for application/json ContentType.
type PostAddEventJSONRequestBody = AddEvent

// PatchCalendarEditCalendarIdJSONRequestBody defines body for PatchCalendarEditCalendarId for application/json ContentType.
type PatchCalendarEditCalendarIdJSONRequestBody = CalendarEdit

// PatchEventEditEventIdJSONRequestBody defines body for PatchEventEditEventId for application/json ContentType.
type PatchEventEditEventIdJSONRequestBody = EventEdit

// PostLoginJSONRequestBody defines body for PostLogin for application/json ContentType.
type PostLoginJSONRequestBody = Login

// PostSignupJSONRequestBody defines body for PostSignup for application/json ContentType.
type PostSignupJSONRequestBody = Signup

// PostTotpAuthenticateJSONRequestBody defines body for PostTotpAuthenticate for application/json ContentType.
type PostTotpAuthenticateJSONRequestBody = EnableTotpVerify

// PatchTotpDisableJSONRequestBody defines body for PatchTotpDisable for application/json ContentType.
type PatchTotpDisableJSONRequestBody = Totp

// PatchTotpEnableVerifyJSONRequestBody defines body for PatchTotpEnableVerify for application/json ContentType.
type PatchTotpEnableVerifyJSONRequestBody = EnableTotpVerify

// PostTotpRecoveryJSONRequestBody defines body for PostTotpRecovery for application/json ContentType.
type PostTotpRecoveryJSONRequestBody = RecoveryCode

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// TODO
	// (POST /addCalendar)
	PostAddCalendar(ctx echo.Context) error
	// TODO
	// (POST /addEvent)
	PostAddEvent(ctx echo.Context) error
	// TODO
	// (DELETE /calendar/delete/{calendar_id})
	DeleteCalendarDeleteCalendarId(ctx echo.Context, calendarId uuid.UUID) error
	// TODO
	// (PATCH /calendar/edit/{calendar_id})
	PatchCalendarEditCalendarId(ctx echo.Context, calendarId uuid.UUID) error
	// TODO
	// (DELETE /event/delete/{event_id})
	DeleteEventDeleteEventId(ctx echo.Context, eventId uuid.UUID) error
	// TODO
	// (PATCH /event/edit/{event_id})
	PatchEventEditEventId(ctx echo.Context, eventId uuid.UUID) error
	// TODO
	// (GET /getCalendars)
	GetGetCalendars(ctx echo.Context) error
	// TODO
	// (GET /getEvents)
	GetGetEvents(ctx echo.Context, params GetGetEventsParams) error
	// TODO
	// (POST /login)
	PostLogin(ctx echo.Context) error
	// TODO
	// (POST /signup)
	PostSignup(ctx echo.Context) error
	// TODO
	// (POST /totp/authenticate)
	PostTotpAuthenticate(ctx echo.Context) error
	// TODO
	// (PATCH /totp/disable)
	PatchTotpDisable(ctx echo.Context) error
	// TODO
	// (POST /totp/enable)
	PostTotpEnable(ctx echo.Context) error
	// TODO
	// (PATCH /totp/enable/verify)
	PatchTotpEnableVerify(ctx echo.Context) error
	// TODO
	// (POST /totp/recovery)
	PostTotpRecovery(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostAddCalendar converts echo context to params.
func (w *ServerInterfaceWrapper) PostAddCalendar(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostAddCalendar(ctx)
	return err
}

// PostAddEvent converts echo context to params.
func (w *ServerInterfaceWrapper) PostAddEvent(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostAddEvent(ctx)
	return err
}

// DeleteCalendarDeleteCalendarId converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteCalendarDeleteCalendarId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "calendar_id" -------------
	var calendarId uuid.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "calendar_id", ctx.Param("calendar_id"), &calendarId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter calendar_id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteCalendarDeleteCalendarId(ctx, calendarId)
	return err
}

// PatchCalendarEditCalendarId converts echo context to params.
func (w *ServerInterfaceWrapper) PatchCalendarEditCalendarId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "calendar_id" -------------
	var calendarId uuid.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "calendar_id", ctx.Param("calendar_id"), &calendarId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter calendar_id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchCalendarEditCalendarId(ctx, calendarId)
	return err
}

// DeleteEventDeleteEventId converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteEventDeleteEventId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "event_id" -------------
	var eventId uuid.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "event_id", ctx.Param("event_id"), &eventId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter event_id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteEventDeleteEventId(ctx, eventId)
	return err
}

// PatchEventEditEventId converts echo context to params.
func (w *ServerInterfaceWrapper) PatchEventEditEventId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "event_id" -------------
	var eventId uuid.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "event_id", ctx.Param("event_id"), &eventId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter event_id: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchEventEditEventId(ctx, eventId)
	return err
}

// GetGetCalendars converts echo context to params.
func (w *ServerInterfaceWrapper) GetGetCalendars(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetGetCalendars(ctx)
	return err
}

// GetGetEvents converts echo context to params.
func (w *ServerInterfaceWrapper) GetGetEvents(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetGetEventsParams
	// ------------- Required query parameter "start_time" -------------

	err = runtime.BindQueryParameter("form", true, true, "start_time", ctx.QueryParams(), &params.StartTime)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter start_time: %s", err))
	}

	// ------------- Required query parameter "end_time" -------------

	err = runtime.BindQueryParameter("form", true, true, "end_time", ctx.QueryParams(), &params.EndTime)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter end_time: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetGetEvents(ctx, params)
	return err
}

// PostLogin converts echo context to params.
func (w *ServerInterfaceWrapper) PostLogin(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostLogin(ctx)
	return err
}

// PostSignup converts echo context to params.
func (w *ServerInterfaceWrapper) PostSignup(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostSignup(ctx)
	return err
}

// PostTotpAuthenticate converts echo context to params.
func (w *ServerInterfaceWrapper) PostTotpAuthenticate(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostTotpAuthenticate(ctx)
	return err
}

// PatchTotpDisable converts echo context to params.
func (w *ServerInterfaceWrapper) PatchTotpDisable(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchTotpDisable(ctx)
	return err
}

// PostTotpEnable converts echo context to params.
func (w *ServerInterfaceWrapper) PostTotpEnable(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostTotpEnable(ctx)
	return err
}

// PatchTotpEnableVerify converts echo context to params.
func (w *ServerInterfaceWrapper) PatchTotpEnableVerify(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PatchTotpEnableVerify(ctx)
	return err
}

// PostTotpRecovery converts echo context to params.
func (w *ServerInterfaceWrapper) PostTotpRecovery(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostTotpRecovery(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/addCalendar", wrapper.PostAddCalendar)
	router.POST(baseURL+"/addEvent", wrapper.PostAddEvent)
	router.DELETE(baseURL+"/calendar/delete/:calendar_id", wrapper.DeleteCalendarDeleteCalendarId)
	router.PATCH(baseURL+"/calendar/edit/:calendar_id", wrapper.PatchCalendarEditCalendarId)
	router.DELETE(baseURL+"/event/delete/:event_id", wrapper.DeleteEventDeleteEventId)
	router.PATCH(baseURL+"/event/edit/:event_id", wrapper.PatchEventEditEventId)
	router.GET(baseURL+"/getCalendars", wrapper.GetGetCalendars)
	router.GET(baseURL+"/getEvents", wrapper.GetGetEvents)
	router.POST(baseURL+"/login", wrapper.PostLogin)
	router.POST(baseURL+"/signup", wrapper.PostSignup)
	router.POST(baseURL+"/totp/authenticate", wrapper.PostTotpAuthenticate)
	router.PATCH(baseURL+"/totp/disable", wrapper.PatchTotpDisable)
	router.POST(baseURL+"/totp/enable", wrapper.PostTotpEnable)
	router.PATCH(baseURL+"/totp/enable/verify", wrapper.PatchTotpEnableVerify)
	router.POST(baseURL+"/totp/recovery", wrapper.PostTotpRecovery)

}