# backend/internal/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Calendar struct {
	ID      uuid.UUID
	OwnerID uuid.UUID
	Name    string
}

type Event struct {
	ID         uuid.UUID
	CalendarID uuid.UUID
	Name       string
	Time       pgtype.Range[pgtype.Timestamptz]
}

type User struct {
	ID           uuid.UUID
	Name         string
	PasswordHash string
	Totp         string
}

type UserRecoveryCode struct {
	ID       int32
	UserID   uuid.UUID
	CodeHash string
	UsedAt   *time.Time
}